requirements.txt
pandas
numpy
matplotlib
torch
transformers
scikit-learn
accelerate
jupyterlab
imbalanced-learn
sentence-transformers
pymongo
pytz
pyYAML
fastdtw
seaborn
setfit